+++ src/NCronJob/Execution/StartupJobManager.cs
@@ -1,13 +1,16 @@
 namespace NCronJob;
 
-internal class StartupJobManager(JobRegistry jobRegistry, JobProcessor jobProcessor)
+internal class StartupJobManager(
+    JobRegistry jobRegistry,
+    JobProcessor jobProcessor,
+    JobExecutionProgressObserver observer)
 {
     private readonly AsyncManualResetEvent startupJobsCompleted = new();
 
     public async Task ProcessStartupJobs(CancellationToken stopToken)
     {
         var startupJobs = jobRegistry.GetAllOneTimeJobs();
-        var startupTasks = startupJobs.Select(definition => CreateExecutionTask(JobRun.Create(definition), stopToken)).ToList();
+        var startupTasks = startupJobs.Select(definition => CreateExecutionTask(JobRun.Create(observer.Report, definition), stopToken)).ToList();
 
         if (startupTasks.Count > 0)
         {